+++ project_config/handle/webpack.js
@@ -3,49 +3,20 @@
 
 const _ = require('lodash');
 
-const makeWebpackDevConfig = require('../util/make_webpack_dev_config');
-const makeWebpackBuildConfig = require('../util/make_webpack_build_config');
-const makeWebpackAnalyzeConfig = require('../util/make_webpack_analyze_config');
+const staticServerUrl = require('../webpack/static_server_url');
+const resolveModules = require('../webpack/resolve_modules');
+const makeDevConfig = require('../webpack/make_dev_config');
+const makeBuildConfig = require('../webpack/make_build_config');
+const makeAnalyzeConfig = require('../webpack/make_analyze_config');
 
 module.exports = (config) => {
     // if is multi modules, no more handling
     if (config.multiModules) return;
 
-    // ''
-    if (!config.staticServerUrl) {
-        config.staticServerDomain = '';
-        config.staticServerDir = '';
-    }
-    else {
-        // http://
-        let hasHttp = _.startsWith(config.staticServerUrl, 'http://');
-        // https://
-        let hasHttps = _.startsWith(config.staticServerUrl, 'https://');
-        // //
-        let hasDoubleSlashes = _.startsWith(config.staticServerUrl, '//');
-        let prefixLength = 0;
+    staticServerUrl(config);
+    resolveModules(config);
 
-        if (hasHttp || hasHttps || hasDoubleSlashes) {
-            hasHttp && (prefixLength = 7);
-            hasHttps && (prefixLength = 8);
-            hasDoubleSlashes && (prefixLength = 2);
-            let threeSlashIndex = config.staticServerUrl.indexOf('/', prefixLength);
-            if (threeSlashIndex < 0) {
-                config.staticServerDomain = config.staticServerUrl;
-                config.staticServerDir = '';
-            }
-            else {
-                config.staticServerDomain = config.staticServerUrl.slice(0, threeSlashIndex);
-                config.staticServerDir = config.staticServerUrl.slice(threeSlashIndex);
-            }
-        }
-        else {
-            config.staticServerDomain = '';
-            config.staticServerDir = config.staticServerUrl;
-        }
-    }
-
-    config.webpackDevConfig = makeWebpackDevConfig(config);
-    config.webpackBuildConfig = makeWebpackBuildConfig(config);
-    config.webpackAnalyzeConfig = makeWebpackAnalyzeConfig(config);
+    config.webpackDevConfig = makeDevConfig(config);
+    config.webpackBuildConfig = makeBuildConfig(config);
+    config.webpackAnalyzeConfig = makeAnalyzeConfig(config);
 };