+++ project_config/webpack/make_analyze_config.js
@@ -3,15 +3,15 @@
 
 var webpack = require('webpack');
 var _ = require('lodash');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
-var getJsEntryPath = require('./get_js_entry_path');
-var makeResolve = require('./webpack/make_resolve');
-var makeHtmlPlugin = require('./webpack/make_html_plugin');
-var makeBabelLoader = require('../../webpack/loaders/babel_loader');
-var makeCssLoader = require('../../webpack/loaders/css_loader');
-var makeLessLoader = require('../../webpack/loaders/less_loader');
-var makeUrlLoader = require('../../webpack/loaders/url_loader');
-var makeHtmlLoader = require('../../webpack/loaders/html_loader');
+var getJsEntryPath = require('./util/get_js_entry_path');
+var makeResolve = require('./util/make_resolve');
+var makeBabelLoader = require('./loaders/babel_loader');
+var makeCssLoader = require('./loaders/css_loader');
+var makeLessLoader = require('./loaders/less_loader');
+var makeUrlLoader = require('./loaders/url_loader');
+var makeHtmlLoader = require('./loaders/html_loader');
 
 module.exports = (config) => {
 
@@ -22,8 +22,9 @@ module.exports = (config) => {
     var htmlLoader = makeHtmlLoader();
 
     var plugins = [
-        new webpack.HotModuleReplacementPlugin(),
-        makeHtmlPlugin(config)
+        new BundleAnalyzerPlugin({
+            analyzerPort: config.analyzerPort
+        })
     ];
 
     // ProvidePlugin
@@ -31,13 +32,12 @@ module.exports = (config) => {
 
     var devConfig = {
         entry: [
-            'webpack-hot-middleware/client?reload=true',
             getJsEntryPath(config)
         ],
         output: {
-            path: config.buildPaths.dev.dir + '/' + config.module + '/',
+            path: config.buildPaths.dev.dir + '/',
             filename: 'index.js',
-            publicPath: config.basePaths.webPrefix + '/dev/' + config.module + '/'
+            publicPath: config.basePaths.webPrefix + '/dev/'
         },
         plugins: plugins,
         module: {